✅ Update tests
Tests updated to work properly after all the module changes
